Adding training logic to core
