+++ configurations/logging.conf
@@ -1,5 +1,5 @@
 [loggers]
-keys=root, gunicorn.error, gunicorn.access, __main__
+keys=root, katana.error, katana.access, __main__
 
 [handlers]
 keys=error_file, syslog, access_file
@@ -11,17 +11,17 @@ keys=generic
 level=INFO
 handlers=error_file, syslog
 
-[logger_gunicorn.error]
+[logger_katana.error]
 level=INFO
 handlers=error_file, syslog
 propagate=0
-qualname=gunicorn.error
+qualname=katana.error
 
-[logger_gunicorn.access]
+[logger_katana.access]
 level=INFO
 handlers=access_file, syslog
 propagate=0
-qualname=gunicorn.access
+qualname=katana.access
 
 [logger___main__]
 level=DEBUG
@@ -32,7 +32,7 @@ qualname=__main__
 [handler_error_file]
 class=logging.handlers.RotatingFileHandler
 formatter=generic
-args=('./logs/gunicorn.error.log', 'a', 10485760)
+args=('./logs/katana.error.log', 'a', 10485760)
 
 [handler_syslog]
 class=StreamHandler
@@ -42,9 +42,9 @@ args=(sys.stdout,)
 [handler_access_file]
 class=logging.handlers.RotatingFileHandler
 formatter=generic
-args=('./logs/gunicorn.access.log', 'a', 10485760)
+args=('./logs/katana.access.log', 'a', 10485760)
 
 [formatter_generic]
 format=%(asctime)s [%(process)d:%(name)s:%(lineno)s] [%(levelname)s] %(message)s
 datefmt=%Y-%m-%d %H:%M:%S
-class=apis.config.ColorFormatter
+class=configurations.config.ColorFormatter
